Add custom reviewboardrc without group set
By unsetting the group, we won't automatically spam the reviewer mailing
list when we first publish commits for review. We can manually add mesos
as the group for the review once we are ready for others to see it.

Make sure and set up the symlink in the 'mesos' directory appropriately
so that this overridden file is used (i.e. after running bootstrap.sh)